Delete unnecessary redux related code. Move lib items to new folder structure.
