+++ app/javascript/App/bundles/NewsfeedPage/NewsfeedPage.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { withRouter, RouteComponentProps } from 'react-router'
 import debounce from 'debounce'
-import { slugify } from '../../lib/utils/slugify'
+import slugify from '~/bundles/common/utils/slugify'
 import LayoutDesktop from '~/bundles/common/components/LayoutDesktop'
 import LayoutTablet from '~/bundles/common/components/LayoutTablet'
 import LayoutMobile from '~/bundles/common/components/LayoutMobile'
@@ -16,7 +16,7 @@ import withDevice from '~/bundles/common/utils/withDevice'
 import EventListener from 'react-event-listener'
 
 import { NewsItem, ContentType, Filters } from './types'
-import { CoinList, CoinWithDetails } from '../common/types'
+import { CoinWithDetails } from '../common/types'
 import {
   getDefaultFilters,
   mergeInitialSocialSourcesForCoinsFilter,
