+++ app/javascript/App/bundles/common/utils/images.ts
