+++ app/javascript/App/bundles/common/utils/url.ts
