--- app/javascript/App/containers/newsfeed/index.js
@@ -1,14 +1,0 @@
-import { createHOC } from '../../lib/redux'
-import actions from './actions'
-import selectors from './selectors'
-import { filterList } from './constants'
-
-export default createHOC({
-  actions,
-  selectors,
-  extraProps: { filterList },
-  onMount: ({ props }) => {
-    props.initializeFilters()
-    props.initializeNewsfeedState()
-  }
-})
