--- app/javascript/App/containers/watchlist/constants.js
@@ -1,1 +1,0 @@
-export const namespace = 'watchlist'
