--- app/javascript/App/containers/watchlist/index.js
@@ -1,11 +1,0 @@
-import { createHOC } from '../../lib/redux'
-import actions from './actions'
-import selectors from './selectors'
-
-export default createHOC({
-  actions,
-  selectors,
-  onMount: (container) => {
-    container.props.fetchCoins()
-  }
-})
