+++ app/javascript/App/normalizers/index.js
@@ -1,9 +1,7 @@
-import newsItems from './newsItems'
 import calendarEvents from './calendarEvents'
 import coins from './coins'
 
 export default {
-  newsItems,
   calendarEvents,
   coins,
 }
