+++ app/javascript/App/withCombinedProviders.tsx
@@ -1,8 +1,6 @@
 import * as React from 'react'
 import CoinListContainer from './bundles/common/containers/CoinListContainer'
 import NewsfeedContainer from '~/bundles/NewsfeedPage/NewsfeedContainer'
-import { MuiThemeProvider } from '@material-ui/core/styles'
-import theme from './theme'
 import { DeviceProvider } from '~/bundles/common/contexts/DeviceContext'
 import { RailsProvider } from '~/bundles/common/contexts/RailsContext'
